+++ src/bigdata_research_tools/workflows/utils.py
@@ -2,35 +2,7 @@
 Script with any common helper functions used across the workflows.
 """
 
-from IPython.display import HTML, display
 from pandas import DataFrame
-
-
-def display_output_chunks_dataframe(final_df):
-    """
-    Display selected document chunks in a formatted HTML view for better readability.
-
-    Args:
-        final_df: DataFrame containing semantic labels with document chunks
-    """
-    output_lines = []
-
-    for row, element in final_df.iterrows():
-        # Add lines to the output list with the company in bold
-        output_lines.append(f"<strong>Company:</strong> {element.Company}<br>")
-        output_lines.append(f"<strong>Sector:</strong> {element.Sector}<br>")
-        output_lines.append(f"<strong>Industry:</strong> {element.Industry}<br>")
-        output_lines.append(f"<strong>Date:</strong> {element.Date}<br>")
-        output_lines.append(f"<strong>Headline:</strong> {element.Headline}<br>")
-        output_lines.append(
-            f"<strong>Sentence Identifier:</strong> {element['Document ID']}<br>"
-        )
-        output_lines.append(f"<strong>Quote:</strong> <em>{element.Quote}</em><br>")
-        output_lines.append(f"<strong>Sub-Theme Label:</strong> {element.Theme}<br>")
-        output_lines.append("--------------------<br>")
-
-    # Join all lines into a single string and display it
-    display(HTML("".join(output_lines)))
 
 
 def get_scored_df(